# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
      "AttemptTrace": {
        "additionalProperties": false,
        "properties": {
          "attempt_index": {
            "minimum": 0.0,
            "title": "Attempt Index",
            "type": "integer"
          },
          "generation": {
            "$ref": "#/components/schemas/GenerationTrace"
          },
          "resources": {
            "$ref": "#/components/schemas/ResourceMetrics"
          },
          "retrieval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalTrace"
              },
              {
                "type": "null"
              }
            ],
            "description": "null for system A."
          },
          "verification": {
            "$ref": "#/components/schemas/VerificationTrace"
          }
        },
        "required": [
          "attempt_index",
          "retrieval",
          "generation",
          "verification",
          "resources"
        ],
        "title": "AttemptTrace",
        "type": "object"
      },
      "CachedRunRef": {
        "additionalProperties": false,
        "properties": {
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "run_id": {
            "maxLength": 120,
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "recorded_at"
        ],
        "title": "CachedRunRef",
        "type": "object"
      },
...
      "ClassificationEvent": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "event": {
            "const": "classification",
            "default": "classification",
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/ClassificationTrace"
          },
          "seq": {
            "description": "0, 1, 2, … within one stream.",
            "minimum": 0,
            "title": "Seq",
            "type": "integer"
          },
          "trace_id": {
            "maxLength": 64,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "seq",
          "payload"
        ],
        "title": "ClassificationEvent",
        "type": "object"
      },
      "ClassificationTrace": {
        "additionalProperties": false,
        "properties": {
          "classifier": {
            "examples": [
              "rule_based"
            ],
            "maxLength": 64,
            "title": "Classifier",
            "type": "string"
          },
          "classifier_confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Classifier Confidence",
            "type": "number"
          },
          "complexity": {
            "$ref": "#/components/schemas/ComplexityLevel"
          },
          "criticality": {
            "$ref": "#/components/schemas/CriticalityLevel"
          },
          "latency_ms": {
            "$ref": "#/components/schemas/Metric"
          },
          "quality_threshold": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Quality Threshold",
            "type": "number"
          },
          "quality_threshold_source": {
            "enum": [
              "classifier",
              "config_default"
            ],
            "title": "Quality Threshold Source",
            "type": "string"
          },
          "reasoning": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "sdlc_stage": {
            "$ref": "#/components/schemas/SDLCStage"
          },
          "security_sensitivity": {
            "$ref": "#/components/schemas/SecuritySensitivity"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "task_type_source": {
            "enum": [
              "classifier",
              "client_hint"
            ],
            "title": "Task Type Source",
            "type": "string"
          }
        },
        "required": [
          "classifier",
          "sdlc_stage",
          "task_type",
          "task_type_source",
          "complexity",
          "criticality",
          "security_sensitivity",
          "quality_threshold",
          "quality_threshold_source",
          "classifier_confidence",
          "latency_ms"
        ],
        "title": "ClassificationTrace",
        "type": "object"
      },
      "ComplexityLevel": {
        "enum": [
          "low",
          "medium",
          "high",
          "very_high",
          "unknown"
        ],
        "title": "ComplexityLevel",
        "type": "string"
      },
      "CriticalityLevel": {
        "description": "Operational criticality of the task outcome.\nHigher criticality → stricter quality threshold, slower escalation path allowed.",
        "enum": [
          "low",
          "medium",
          "high",
          "critical",
          "unknown"
        ],
        "title": "CriticalityLevel",
        "type": "string"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "description": "Human-readable; never contains stack traces, paths or config.",
            "maxLength": 500,
            "title": "Message",
            "type": "string"
          },
          "retry_after_s": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After S"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ErrorEvent": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "event": {
            "const": "error",
            "default": "error",
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "seq": {
            "description": "0, 1, 2, … within one stream.",
            "minimum": 0,
            "title": "Seq",
            "type": "integer"
          },
          "trace_id": {
            "maxLength": 64,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "seq",
          "payload"
        ],
        "title": "ErrorEvent",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "request_id": {
            "maxLength": 64,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EscalationEvent": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "event": {
            "const": "escalation",
            "default": "escalation",
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/EscalationStep"
          },
          "seq": {
            "description": "0, 1, 2, … within one stream.",
            "minimum": 0,
            "title": "Seq",
            "type": "integer"
          },
          "trace_id": {
            "maxLength": 64,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "seq",
          "payload"
        ],
        "title": "EscalationEvent",
        "type": "object"
      },
...
      "EscalationSummary": {
        "additionalProperties": false,
        "properties": {
          "exhausted": {
            "description": "True if every allowed escalation failed.",
            "title": "Exhausted",
            "type": "boolean"
          },
          "max_escalations": {
            "minimum": 0.0,
            "title": "Max Escalations",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/EscalationStep"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "max_escalations",
          "exhausted"
        ],
        "title": "EscalationSummary",
        "type": "object"
      },
      "ExecutionContext": {
        "additionalProperties": false,
        "properties": {
          "cached_from": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CachedRunRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "generation_source": {
            "$ref": "#/components/schemas/GenerationSource"
          },
          "producer": {
            "$ref": "#/components/schemas/MachineInfo"
          },
          "research_evidence": {
            "description": "True only if the answer came from a real model run. Always false for mock output. Does NOT mean the trace belongs to any benchmark.",
            "title": "Research Evidence",
            "type": "boolean"
          },
          "served_by_machine_id": {
            "pattern": "^[a-z0-9][a-z0-9-]{0,62}$",
            "title": "Served By Machine Id",
            "type": "string"
          }
        },
        "required": [
          "producer",
          "served_by_machine_id",
          "generation_source",
          "research_evidence"
        ],
        "title": "ExecutionContext",
        "type": "object"
      },
      "GenerationDeltaEvent": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "event": {
            "const": "generation.delta",
            "default": "generation.delta",
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/GenerationDeltaPayload"
          },
          "seq": {
            "description": "0, 1, 2, … within one stream.",
            "minimum": 0,
            "title": "Seq",
            "type": "integer"
          },
          "trace_id": {
            "maxLength": 64,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "seq",
          "payload"
        ],
        "title": "GenerationDeltaEvent",
        "type": "object"
      },
...
      "GenerationEvent": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "event": {
            "const": "generation",
            "default": "generation",
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/GenerationTrace"
          },
          "seq": {
            "description": "0, 1, 2, … within one stream.",
            "minimum": 0,
            "title": "Seq",
            "type": "integer"
          },
          "trace_id": {
            "maxLength": 64,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "seq",
          "payload"
        ],
        "title": "GenerationEvent",
        "type": "object"
      },
...
      "Outcome": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "description": "Final answer, or the INSUFFICIENT EVIDENCE refusal text.",
            "title": "Answer",
            "type": "string"
          },
          "answer_attempt_index": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Attempt whose answer is returned; null on refusal.",
            "title": "Answer Attempt Index"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          }
        },
        "required": [
          "verdict",
          "answer"
        ],
        "title": "Outcome",
        "type": "object"
      },
...
      "PipelineTrace": {
        "additionalProperties": false,
        "description": "The full trace of one query through the EIH pipeline.",
        "properties": {
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/AttemptTrace"
            },
            "minItems": 1,
            "title": "Attempts",
            "type": "array"
          },
          "classification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClassificationTrace"
              },
              {
                "type": "null"
              }
            ],
            "description": "null if the system does not classify (A, B)."
          },
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "escalation": {
            "$ref": "#/components/schemas/EscalationSummary"
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionContext"
          },
          "outcome": {
            "$ref": "#/components/schemas/Outcome"
          },
          "query": {
            "maxLength": 2000,
            "title": "Query",
            "type": "string"
          },
          "request_id": {
            "maxLength": 64,
            "title": "Request Id",
            "type": "string"
          },
          "system": {
            "$ref": "#/components/schemas/SystemId"
          },
          "totals": {
            "$ref": "#/components/schemas/ResourceMetrics"
          },
          "trace_id": {
            "maxLength": 64,
            "title": "Trace Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "trace_id",
          "request_id",
          "created_at",
          "query",
          "system",
          "execution",
          "classification",
          "attempts",
          "escalation",
          "outcome",
          "totals"
        ],
        "title": "PipelineTrace",
        "type": "object"
      },
...
      "ResourceMetrics": {
        "additionalProperties": false,
        "description": "Resource accounting for one attempt, or totals for the whole trace.\n\nTotals inherit the weakest tier of their components (MEASURED is\nstrongest); see docs/API_CONTRACT.md §5.",
        "properties": {
          "co2e_grams": {
            "$ref": "#/components/schemas/Metric"
          },
          "cost_usd": {
            "$ref": "#/components/schemas/Metric"
          },
          "energy_cpu_joules": {
            "$ref": "#/components/schemas/Metric"
          },
          "energy_gpu_joules": {
            "$ref": "#/components/schemas/Metric"
          },
          "energy_joules": {
            "$ref": "#/components/schemas/Metric",
            "description": "cpu + gpu + rerank energy."
          },
          "energy_rerank_joules": {
            "$ref": "#/components/schemas/Metric"
          },
          "input_tokens": {
            "$ref": "#/components/schemas/Metric"
          },
          "latency_ms": {
            "$ref": "#/components/schemas/Metric"
          },
          "output_tokens": {
            "$ref": "#/components/schemas/Metric"
          }
        },
        "required": [
          "latency_ms",
          "input_tokens",
          "output_tokens",
          "cost_usd",
          "energy_joules",
          "energy_cpu_joules",
          "energy_gpu_joules",
          "energy_rerank_joules",
          "co2e_grams"
        ],
        "title": "ResourceMetrics",
        "type": "object"
      },
      "RetrievalEvent": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "event": {
            "const": "retrieval",
            "default": "retrieval",
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/RetrievalTrace"
          },
          "seq": {
            "description": "0, 1, 2, … within one stream.",
            "minimum": 0,
            "title": "Seq",
            "type": "integer"
          },
          "trace_id": {
            "maxLength": 64,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "seq",
          "payload"
        ],
        "title": "RetrievalEvent",
        "type": "object"
      },
...
      "SDLCStage": {
        "description": "Software Development Life Cycle stage.",
        "enum": [
          "requirements",
          "architecture",
          "development",
          "testing",
          "code_review",
          "deployment",
          "operations",
          "maintenance",
          "unknown"
        ],
        "title": "SDLCStage",
        "type": "string"
      },
      "SecuritySensitivity": {
        "enum": [
          "none",
          "low",
          "medium",
          "high",
          "unknown"
        ],
        "title": "SecuritySensitivity",
        "type": "string"
      },
...
      "SystemId": {
        "description": "Ablation ladder position (CLAUDE.md, 'What this project is').",
        "enum": [
          "A",
          "B",
          "C",
          "D",
          "E"
        ],
        "title": "SystemId",
        "type": "string"
      },
      "TaskType": {
        "description": "Fine-grained task type within an SDLC stage.\nInitial supported set (Phase-0 definition); extend in Phase-1.",
        "enum": [
          "requirement_understanding",
          "requirement_retrieval",
          "architecture_qa",
          "dependency_understanding",
          "architecture_decision_support",
          "code_explanation",
          "code_generation",
          "repository_assistance",
          "test_generation",
          "test_explanation",
          "test_failure_analysis",
          "defect_detection",
          "risk_identification",
          "review_assistance",
          "change_impact_analysis",
          "dependency_analysis",
          "error_analysis",
          "incident_retrieval",
          "root_cause_assistance",
          "technical_debt_analysis",
          "change_understanding",
          "historical_reasoning",
          "unknown"
        ],
        "title": "TaskType",
        "type": "string"
      },
      "TraceCompletedEvent": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "event": {
            "const": "trace.completed",
            "default": "trace.completed",
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/PipelineTrace"
          },
          "seq": {
            "description": "0, 1, 2, … within one stream.",
            "minimum": 0,
            "title": "Seq",
            "type": "integer"
          },
          "trace_id": {
            "maxLength": 64,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "seq",
          "payload"
        ],
        "title": "TraceCompletedEvent",
        "type": "object"
      },
      "TraceEvent": {
        "discriminator": {
          "mapping": {
            "classification": "#/components/schemas/ClassificationEvent",
            "error": "#/components/schemas/ErrorEvent",
            "escalation": "#/components/schemas/EscalationEvent",
            "generation": "#/components/schemas/GenerationEvent",
            "generation.delta": "#/components/schemas/GenerationDeltaEvent",
            "retrieval": "#/components/schemas/RetrievalEvent",
            "trace.completed": "#/components/schemas/TraceCompletedEvent",
            "trace.started": "#/components/schemas/TraceStartedEvent",
            "verification": "#/components/schemas/VerificationEvent"
          },
          "propertyName": "event"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/TraceStartedEvent"
          },
          {
            "$ref": "#/components/schemas/ClassificationEvent"
          },
          {
            "$ref": "#/components/schemas/RetrievalEvent"
          },
          {
            "$ref": "#/components/schemas/GenerationDeltaEvent"
          },
          {
            "$ref": "#/components/schemas/GenerationEvent"
          },
          {
            "$ref": "#/components/schemas/VerificationEvent"
          },
          {
            "$ref": "#/components/schemas/EscalationEvent"
          },
          {
            "$ref": "#/components/schemas/TraceCompletedEvent"
          },
          {
            "$ref": "#/components/schemas/ErrorEvent"
          }
        ]
      },
      "TraceRequest": {
        "additionalProperties": false,
        "properties": {
          "query": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Query",
            "type": "string"
          },
          "repository": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9_.-]{1,100}/[A-Za-z0-9_.-]{1,100}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restrict retrieval to one ingested repository (owner/name).",
            "title": "Repository"
          },
          "system": {
            "$ref": "#/components/schemas/SystemId",
            "default": "E"
          },
          "task_type_hint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "query"
        ],
        "title": "TraceRequest",
        "type": "object"
      },
      "TraceStartedEvent": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "event": {
            "const": "trace.started",
            "default": "trace.started",
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/TraceStartedPayload"
          },
          "seq": {
            "description": "0, 1, 2, … within one stream.",
            "minimum": 0,
            "title": "Seq",
            "type": "integer"
          },
          "trace_id": {
            "maxLength": 64,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "seq",
          "payload"
        ],
        "title": "TraceStartedEvent",
        "type": "object"
      },
...
      "VerificationEvent": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contract Version",
            "type": "string"
          },
          "event": {
            "const": "verification",
            "default": "verification",
            "title": "Event",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/VerificationTrace"
          },
          "seq": {
            "description": "0, 1, 2, … within one stream.",
            "minimum": 0,
            "title": "Seq",
            "type": "integer"
          },
          "trace_id": {
            "maxLength": 64,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "seq",
          "payload"
        ],
        "title": "VerificationEvent",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Frozen v1 contract between the EIH research pipeline and the web platform. See docs/API_CONTRACT.md.",
    "title": "EIH Pipeline Trace API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/trace": {
      "post": {
        "operationId": "trace_v1_trace_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineTrace"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Pipeline and cache both unavailable"
          }
        },
        "summary": "Run one query and return the full pipeline trace",
        "tags": [
          "Trace v1"
        ]
      }
    },
    "/v1/trace/stream": {
      "post": {
        "operationId": "trace_stream_v1_trace_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/TraceEvent"
                }
              }
            },
            "description": "Server-Sent Events. Each event's `event:` field equals the JSON `event` discriminator; `data:` is one TraceEvent."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited; see Retry-After"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Pipeline and cache both unavailable"
          }
        },
        "summary": "Run one query and stream trace events as they happen",
        "tags": [
          "Trace v1"
        ]
      }
    }
  }
}